requirements.txt
streamlit>=1.37
openai>=1.40
python-dotenv>=1.0
matplotlib>=3.8
numpy>=1.26
reportlab>=4.2
pandas>=2.2
plotly>=5.24
